chore: rename lib/deskflow/mouse_types => lib/deskflow/MouseTypes
