+++ src/lib/deskflow/IClient.h
@@ -10,7 +10,7 @@
 #include "deskflow/ClipboardTypes.h"
 #include "deskflow/IScreen.h"
 #include "deskflow/KeyTypes.h"
-#include "deskflow/mouse_types.h"
+#include "deskflow/MouseTypes.h"
 #include "deskflow/option_types.h"
 
 #include <string>
